Activate context menu
Thanks Janus2 for the hint to SetSelectionMessage().

At least one issue left:
When there are multiple rows selected, a right-click will change
the selection to the one clicked on. Not sure if that can be dealt
with, as the BColumnListView doesn't have a MouseDown() where we
could check if it's a right-click, but the SelectionMessage is
sent right away and the de-selecting is done without our doing.

Work around (almost an easter-egg...):
When selecting multiple rows, keep holding that SHIFT key until
after you do did the right-click (or to drag'n'drop to download).